moved index file to separate route,added extra data and filter out weekly reports
