+++ src/pages/blog/markets-news/index.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import MarketNews from './_markets-news'
+import MarketNews from '../_markets-news'
 import Layout from 'components/layout/layout'
 import { WithIntl } from 'components/localization'
